requirements.txt
playwright==1.58.0
beautifulsoup4==4.12.3
google-auth==2.29.0
oauth2client==4.1.3
python-dotenv==1.2.2
gspread==6.2.1
